+++ lib/modules/test/test_service.dart
@@ -1,0 +1,9 @@
+﻿// Module: test
+// Updated at: 2026-04-02 12:13:35
+// Commit: chore(test): update dependencies and cleanup unused imports in test
+
+class testComponentState {
+  final String commitRef = '51';
+  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
+  final DateTime timestamp = DateTime.parse('2026-04-02T12:13:35+05:00');
+}